academeet/requirements.txt
Django==5.2.6
dj-database-url
python-dotenv
djangorestframework
Pillow


